requirements.txt
dspy-ai>=2.4.0
datasets>=2.0.0
pandas>=1.5.0
Pillow>=9.0.0
numpy>=1.21.0
nltk>=3.7
rouge-score>=0.1.2
jiwer>=2.0.0
jieba>=0.42.1
pycocoevalcap>=1.2
torch>=1.12.0
transformers>=4.20.0 